% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ATE_from_simulated_data}
\alias{ATE_from_simulated_data}
\title{Calculate ATE from simulated data}
\usage{
ATE_from_simulated_data(from_simulate_data)
}
\arguments{
\item{from_simulate_data}{the object returned by ATEBounds::simulate_data}
}
\value{
a numeric value giving the ATE = P(Y = 1 | X = 1, U) - P(Y = 1 | X = 0, U). To make sure this
is accurate, use large n in your ATEBounds::simulate_data call.
}
\description{
Calculates the ATE based on the results from ATEBounds::simulate_data.
}
