% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpbounds_misc.R
\name{bpbounds_from_data}
\alias{bpbounds_from_data}
\title{Wrapper of bpbounds::bpbounds}
\usage{
bpbounds_from_data(input_data, Y = "Y", X = "X", Z = NULL, ...)
}
\arguments{
\item{input_data}{data.frame with at least three columns containing actual data.}

\item{Y}{column to be used for output}

\item{X}{column to be used for exposure}

\item{Z}{columns to be used for IVs. If NULL, all columns with names matching Z\link{0-9}+ will be used.}
}
\description{
Function that allows you to get bounds using the bpbounds::bpbounds
function from actual data.
}
