% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymake_helpers.R
\name{create_script}
\alias{create_script}
\title{Create Polymake Script}
\usage{
create_script(
  vertices,
  output_file_name,
  polymake_script_name = "polymake_script",
  overwrite = FALSE,
  output_folder = "."
)
}
\arguments{
\item{vertices}{output from \code{create_vertices}}

\item{output_file_name}{name to be used for polymake output}

\item{polymake_script_name}{name to be used for polymake script}

\item{overwrite}{logical; if true, existing files will be overwritten}

\item{output_folder}{folder where you want polymake script and output to be saved}
}
\value{
Nothing is returned, but a file is written that can be fed to polymake.
}
\description{
Function that writes script to run through polymake.
}
