% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymake_helpers.R
\name{create_vertices}
\alias{create_vertices}
\title{Create vertices for polymake program}
\usage{
create_vertices(
  n_z_levels,
  data_format = c("bivariate", "trivariate"),
  x_monotone = FALSE,
  y_monotone = FALSE,
  include_alpha = TRUE
)
}
\arguments{
\item{n_z_levels}{number of categories for IV}

\item{data_format}{one of 'bivariate' or 'trivariate'. If bivariate, output will give extremes in terms of P(Y|Z) and P(X|Z).
If trivariate, output will give extremes in therms of P(Y,X|Z).}

\item{x_monotone}{logical; is P(X = 1 | Z = z) monotonically increasing?}

\item{y_monotone}{logical; is P(Y = 1 | Z = z) monotonically increasing?}

\item{include_alpha}{logical; should a column for alpha = P(Y = 1 | X = 1, U) - P(Y = 1 | X = 0, U) be included?}
}
\value{
matrix of the extreme vertices in the specified scenario.
}
\description{
Function to generate vertices that can be fed to the \href{https://www.polymake.org}{polymake} program for a number of different scenarios:
}
\details{
\itemize{
\item IV with a user specified number of categories
\item \eqn{P(X = 1 | Z = z)} monotonically increasing
\item \eqn{P(Y = 1 | Z = z)} monotonically increasing
\item include column for \eqn{\alpha} (the ACE)
\item output in terms of bivariate probabilities (\eqn{P(Y = y | Z = z)} and \eqn{P(X = x | Z = z)}), or trivariate probabilities (\eqn{P(Y = y, X = x | Z = z)}).
}
}
