% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bounds.R
\name{get_bounds}
\alias{get_bounds}
\title{Function to return ACE bounds from P(X = 1 | Z), P(Y = 1 | Z), or P(X,Y | Z).}
\usage{
get_bounds(
  gammas = NULL,
  thetas = NULL,
  zetas = NULL,
  stop = TRUE,
  warning = TRUE,
  from_polymake = NULL,
  x_mono = FALSE,
  y_mono = FALSE
)
}
\arguments{
\item{gammas}{vector of length \eqn{k} giving \eqn{P(Y = 1 | Z = z), z = 0, ..., (k - 1)}}

\item{thetas}{vector of length \eqn{k} giving \eqn{P(X = 1 | Z = z), z = 0, ..., (k - 1)}}

\item{zetas}{array of dimensions \eqn{2\cdot 2\cdot k} giving P(Y = y, X = x | Z = z) at position (y+1,x+1,z+1),}

\item{stop}{logical; should function stop with error if constraints are not met?}

\item{warning}{logical; if stop is \code{FALSE}, this specifies whether a warning should be returned or not. Default: \code{TRUE}}

\item{from_polymake}{result of \code{read_polymake_results}. If \code{NULL} (default), \code{x_mono} and \code{y_mono} must be specified,
and previously created matrices will be used.}

\item{x_mono}{logical; should P(X = 1 | Z = z) be assumed monotonically increasing? (Default: FALSE)}

\item{y_mono}{logical; should P(Y = 1 | Z = z) be assumed monotonically increasing? (Default: FALSE)}
}
\value{
A list with the following elements:
\itemize{
\item \code{lower_bounds}: a vector of values that the lower bound must be greater than
\item \code{upper_bounds}: a vector of values that the upper bound must be smaller than
\item \code{interval}: a vector with the lower bound (\code{max(lower_bounds)}) and upper bound (\code{min(upper_bounds)}).
\item \code{width}: simply the width of the interval
\item \code{constraints_violated}: a logical value that indicates if any constraints are violated (\code{TRUE}) or not (\code{FALSE})
\item \code{assumptions}: logical vector of length two giving \code{x_mono} and \code{y_mono}. For bookkeeping.
}
}
\description{
Using the \href{https://www.polymake.org}{polymake} program, a matrix \eqn{A} is created such that \eqn{A \cdot p} will have non-negative
elements if the probabilities \eqn{p} specify a probability distribution that does not violate the IV model. \eqn{p} can be probabilities
of either P(X = x | Z = z) and P(Y = y | Z = z), or P(Y = y, X = x | Z = z).
}
