% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_joint.R
\name{joint_from_marginal}
\alias{joint_from_marginal}
\title{P(X,Y|Z) from P(X|Z), P(Y|Z), and Cov(X,Y|Z)}
\usage{
joint_from_marginal(
  thetas,
  gammas,
  cov,
  stop = TRUE,
  message = TRUE,
  x_mono = FALSE,
  y_mono = FALSE,
  from_polymake = NULL,
  return_bounds = FALSE
)
}
\arguments{
\item{thetas}{vector of length \code{n_z_levels} giving P(X = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}

\item{gammas}{vector of length \code{n_z_levels} giving P(Y = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}

\item{cov}{vector of length \code{n_z_levels} giving Cov(X, Y | Z = z), z = 0,..., (\code{n_z_levels} - 1).}

\item{stop}{logical; if \code{TRUE} (default), an error is thrown if the resulting probability distribution
is invalid, or does not satisfy the IV inequalities. Otherwise, a warning is printed, and all
probabilities returned set to NA.}

\item{message}{logical; only used when \code{stop = TRUE}. If false, no warning message is printed when the
resulting probability distribution is invalid. (Default: \code{TRUE})}

\item{from_polymake}{matrix created by the polymake program with the constraints a valid joint distribution
should meet. If \code{NULL} (default), a previously prepared matrix is used if data seems to match one of these.}

\item{return_bounds}{logical; should the resulting bounds be included in the output?}

\item{mono_x}{logical; should \eqn{P(X = 1 | Z = z)} be monotonically increasing?}

\item{mono_y}{logical; should \eqn{P(Y = 1 | Z = z)} be monotonically increasing?}
}
\value{
A \eqn{2x2xk} array with \eqn{P(Y = y, X = x | Z = z)} as entry (y+1, x+1, z+1).
}
\description{
For specified values of \eqn{P(X = 1 | Z = z)}, \eqn{P(Y = 1 | Z = z)}, and \eqn{\text{Cov}(X, Y | Z = z)}, \eqn{z = 0,...,k-1},
the joint \eqn{P(Y = y, X = x | Z = z)} is reconstructed.
}
