% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals_from_logit.R
\name{marginals_from_logit}
\alias{marginals_from_logit}
\title{Get P(X = 1 | Z_j), P(X) from a specified logistic model}
\usage{
marginals_from_logit(b0, b, p.z = rep(list(c(1, 1, 1)/3), 3), nSim = 10000)
}
\arguments{
\item{b0}{numeric; value of intercept \eqn{\beta_0}.}

\item{b}{a numeric vector of slope coefficients \eqn{\beta_1, ..., \beta_n} in logit probability model. Length of vector specifies number of IVs.}

\item{p.z}{a list of vectors of probabilities of \eqn{P(Z_j = z), z = 0,1,2,...,k}. Length of list must be same as
length of \code{b}, and each element of of the list must have length \eqn{l}, number of categories of the IVs.}

\item{nSim}{number of simulations to use when using stochastic integration}
}
\value{
tibble with one row for each IV, and a total of k+2 columns: one column to id the IVs (\code{j}), one column for each \eqn{z} giving
\eqn{P(X = 1 | Z_j = z)}, and one column giving \eqn{P(X = 1)}.
}
\description{
This code calculates \eqn{P(X = 1 | Z_j = z)}, \eqn{z = 0,1,2,...,k}, for any number of \eqn{Z_j}'s. Assuming
\eqn{P(X = 1 | Z_1 = z_1, ..., Z_n = z_n) = \text{expit}(\beta_0 + \sum_j \beta_j z_j)}, this code numerical integrates
out all but one IV for all IVs. Here, it is assumed that all IVs are independent. Values of \eqn{\beta}'s and \eqn{P(Z_j = z), z=0,1,...,k}
are user specified.
}
\details{
This code is fast if all \eqn{\left\{P(Z_j = z)\right\}_{z=0}^k} are the same for all \eqn{j}, and all coefficients \eqn{\beta_j}'s are the same.
In this case, it can handle an arbitrary number of IVs.

For other cases, it will brute force the integral to compute exactly when possible. When number of IVs is greater than 15, it will
switch to stochastic integration (essentially MCMC).
}
