% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymake_helpers.R
\name{perform_transformation}
\alias{perform_transformation}
\title{Function to transform etas and deltas to gammas and thetas}
\usage{
perform_transformation(
  etas,
  deltas,
  data_format = c("bivariate", "trivariate"),
  include_alpha = FALSE
)
}
\arguments{
\item{etas}{vector of length 2 giving values of P(Y = 1 | X = 0, U), P(Y = 1 | X = 1, U)}

\item{deltas}{vector of same length as number of categories for IV giving P(X = 1 | Z = z, U), z = 0, ..., k, k = number of IV categories minus 1.}

\item{data_format}{one of 'bivariate' or 'trivariate'. If bivariate, output will give extremes in terms of P(Y|Z) and P(X|Z).
If trivariate, output will give extremes in therms of zeta{yxz} = P(Y = y, X = x | Z = z).}

\item{include_alpha}{logical; whether to include column for alpha or not.}
}
\description{
This is a helper function that allows us to transform extreme values of \eqn{P(Y = 1 | X = x, U)}, \eqn{P(X = 1 | Z = z, U)} into
values of \eqn{P(Y = y | Z = z, U)} and \eqn{P(X = x | Z = z, U)}, or \eqn{P(Y = y, X = x | Z = z, U)}. It is used to programmatically
generate vertices needed in the \href{https://www.polymake.org}{polymake} program.
}
