% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_joint.R
\name{plot_valid_covs}
\alias{plot_valid_covs}
\title{Plot valid covs}
\usage{
plot_valid_covs(pot_covs, Zs = c(0, 1))
}
\arguments{
\item{pot_covs}{output from \code{potential_covs()}}
}
\description{
Visualize the area of valid covariances.
}
