% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_joint.R
\name{potential_covs}
\alias{potential_covs}
\title{Valid range of Covs}
\usage{
potential_covs(thetas, gammas, x_mono = FALSE)
}
\arguments{
\item{thetas}{vector of length \code{n_z_levels} giving P(X = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}

\item{gammas}{vector of length \code{n_z_levels} giving P(Y = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}

\item{x_mono}{logical; should monotonicity of P(X = 1 | Z = z) be assumed?}
}
\value{
A list with four elements:
\itemize{
\item \code{thetas} and \code{gammas} provided
\item \code{potential_cov_range} is a tibble with three columns: \code{min} and \code{max} give bounds on Cov(X,Y|Z=z) for the
value of \code{z}
\item \code{constraints}is a tibble with four columns: \code{min} and \code{max} give bounds on Cov(X,Y|Z = z1) - Cov(X,Y | Z = z2) for the
corresponding values of \code{z1} and \code{z2}.
}
}
\description{
Given values of P(X = 1 | Z = z) and P(Y = 1 | Z = z) (for all z = 0, ..., \code{n_z_levels} - 1),
this function returns constraints that the quantities Cov(X,Y|Z = z) and Cov(X,Y|Z = z1) - Cov(X,Y|Z = z2)
must satisfy for the reconstructed distribution of $P(X = x, Y = y | Z = z)$ to be a valid
probability distribution and not violate the IV inequalities. Note: fulfilling these constraints is a
necessary but not sufficient requirement for the reconstructed distribution to be valid and not violate the IV
inequalities.
}
