% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_manipulations.R
\name{probs_from_data}
\alias{probs_from_data}
\title{Calculate gammas/thetas/zetas from full data}
\usage{
probs_from_data(dat, X, Y, Z, data_format)
}
\arguments{
\item{dat}{data.frame with columns containing observations of X,Y,Z}

\item{X}{variable with observations on X}

\item{Y}{variable with observations on Y}

\item{Z}{variable with observations on Z}

\item{data_format}{bivariate if you want P(X = 1 | Z = z) and P(Y = 1 | Z = z), trivariate if you want P(X = x, Y = y | Z = z).
If \code{NULL} (default), trivariate is chosen if both X and Y are}
}
\value{
a list with the following elements:
\itemize{
\item if \code{data_format} is bivariate, a vector \code{gammas} of length number of different Z values giving P(Y = 1 | Z = z) is returned if \code{Y} specified, and \code{thetas} of length number of different Z values giving P(X = 1 | Z = z) if \code{X} specified.
\item if \code{data_format} is trivariate, a 2x2xn_z_leveles array \code{zetas} giving P(Y = y, X = x | Z = z) as entry (x+1, y+1, z+1).
}
}
\description{
Calculates either P(Y = 1 | Z = z) (\code{gammas}) and P(X = 1 | Z = z) (\code{thetas}), or P(X = x, Y = y | Z = z) (\code{zetas}) from data.frame/tibble with columns X,Y,Z
}
