% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rbivNorm}
\alias{rbivNorm}
\title{Simulate dependent categorical variables}
\usage{
rbivNorm(n, mu = 0, var1 = 1, var2 = 1, rho = 0, probs = c(0.25, 0.5, 0.25))
}
\description{
Simulate dependent categorical variables
}
