% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymake_helpers.R
\name{read_polymake_results}
\alias{read_polymake_results}
\title{Read output from polymake}
\usage{
read_polymake_results(output_file, data_format = c("bivariate", "trivariate"))
}
\arguments{
\item{output_file}{path to output file from polymake script. \code{output_file} output from \code{create_script}}

\item{data_format}{one of 'bivaraite' or 'trivariate'. If bivariate, output will give extremes in terms of P(Y|Z) and P(X|Z).
If trivariate, output will give extremes in terms of P(Y,X|Z).}
}
\value{
tibble with matrix of constraints for bounds. I.e. \eqn{A \cdot p} must be non-negative for \eqn{p} to satisfy the IV model.
}
\description{
This function reads in the output generated from running the polymake script created by \code{create_script}.
}
