% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_joint.R
\name{sample_joint_probs}
\alias{sample_joint_probs}
\title{Sample valid distributions P(X,Y|Z)}
\usage{
sample_joint_probs(
  pot_covs,
  n = 1000,
  max_rejections = Inf,
  x_mono = FALSE,
  y_mono = FALSE,
  from_polymake = NULL,
  return_bounds = FALSE,
  print_progress = TRUE,
  print_as_progress = NULL
)
}
\arguments{
\item{pot_covs}{output from \code{potential_covs()}}

\item{n}{number of samples to draw}

\item{max_rejections}{number of rejections at which point we move on}

\item{x_mono}{should it be assumed that P(X = 1 | Z = z) is monotonically increasing?}

\item{y_mono}{should it be assumed that P(Y = 1 | Z = z) is monotonically increasing?}

\item{from_polymake}{if NULL (default), we look through the pre-constructed matrices to find one that matches our setup. Otherwise, a matrix can be provided.}

\item{return_bounds}{logical; should the bounds be returned for each sampled joint distribution? (Default: FALSE)}

\item{print_progress}{logical; should process be printed?}

\item{print_as_progress}{if not NULL (default), whatever is provided will be printed in beginning. Useful for batch runs.}
}
\value{
\verb{n x 4} tibble with the following columns:
\itemize{
\item \code{id}: simply ids the run, i.e. is \code{1:n}
\item \code{covs}: list of vectors where each vector gives the sampled Cov(X,Y|Z = z) for each z.
\item \code{joint}: list of \verb{2 x 2 x k} arrays giving the joint conditional distribution reconstructed using
the drawn covariances
\item \code{n_rejected}: number of draws that were rejected due to the reconstructed conditional distribution
not satisfying the constraints.
}
}
\description{
Returns a \verb{n x 4} tibble with one id column, and one column for each level
of Z containing a random draw of a potential value of Cov(X,Y|Z=z). These values are drawn such that
the joint conditional distribution P(X,Y|Z) that can be constructed from these and the values of
P(X = 1 | Z) and P(Y = 1 | Z) that were provided to \code{potential_covs} satisfy the IV inequalities.
}
