% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data from IV model}
\usage{
simulate_data(
  sample_size = 10000,
  IVs_ps = list(c(0.5, 0.5), c(0.5, 0.5)),
  indIVs_on_X = NULL,
  indIVs_on_Y = NULL,
  U = distributions3::Bernoulli(0.5),
  X_intercept = 0.05,
  U_on_X = 0.1,
  Y_intercept = 0.05,
  X_on_Y = 0.05,
  U_on_Y = 0.1,
  rho = NULL,
  depIVs_on_X = NULL,
  dep_probs = c(0.5, 0.5),
  depIVs_on_Y = NULL
)
}
\arguments{
\item{sample_size}{default 10000; number of observations to simulate.}

\item{IVs_ps}{list of numeric vectors. Length of list specifies number of independent IVs to include. Each vector
gives probabilities of \eqn{P(Z = i)}, and must therefore sum to 1.}

\item{indIVs_on_X}{numeric vector specifying the effects of the independent IVs on X (i.e. \eqn{\alpha_1, ..., \alpha_n}).}

\item{indIVs_on_Y}{numeric vector specifying the effects of the independent IVs on Y (i.e. \eqn{\beta_1, ..., \beta_n}).}

\item{U}{\code{distributions3} object specifying the distribution to be used for the unmeasured confounder. Default is \code{Bernoulli(0.5)}.}

\item{X_intercept}{numeric value to use for \eqn{\gamma_x}. Default: 0.05.}

\item{U_on_X}{numeric value to use for \eqn{\alpha_U}. Default: 0.1.}

\item{Y_intercept}{numeric value to use for \eqn{\gamma_y}. Default: 0.05.}

\item{X_on_Y}{numeric value to use for \eqn{\beta_X}. Default: 0.05.}

\item{U_on_Y}{numeric value to use for \eqn{\beta_U}. Default: 0.1.}

\item{rho}{numeric value specifying \eqn{\rho}. Default: NULL.}

\item{depIVs_on_X}{numeric vector specifying the effects of the dependent IVs on X (i.e. \eqn{\alpha_{D1}} and \eqn{\alpha_{D2}}).}

\item{dep_probs}{numeric vector. The length specifies the number of categories for the IVs, and the values give \eqn{P(Z_i = z)}.}

\item{depIVs_on_Y}{numeric vector specifying the effects of the dependent IVs on Y (i.e. \eqn{\beta_{D1}} and \eqn{\beta_{D2}}).}
}
\description{
This function allows the user to simulate data from an IV model
with an arbitrary number of IVs with the option to include IVs
that are dependent on each other, and IVs with a direct effect
on the outcome.
}
\details{
Data are simulated from a logistic model as follows:
\itemize{
\item A binary confounder \eqn{U} is simulated with user specified \eqn{P(U = 1)}.
\item A user specified number of independent IVs (\eqn{Z_1, ..., Z_n}) that are either
di- or trichotomous. Probabilities \eqn{P(Z_j = z), z=0,1} (and \eqn{P(Z_j = 2)} if trichotomous
case is considered) are specified by the user.
\item Optionally, a pair of dependent IVs, \eqn{Z_{D1}, Z_{D2}}, simulated as follows:
\itemize{
\item User specifies \eqn{\rho}
\item \eqn{(N_{D1}, N_{D2}) \sim N\left(\begin{pmatrix} 0 \\ 0 \end{pmatrix},\begin{pmatrix} 1 & \rho \\ \rho & 1 \end{pmatrix} \right)},
then \deqn{
          Z_{Di} = \left\{\begin{array}{cl} 0 & \text{ if } N_{Di} \le z_{p_1} \\ 1 & \text{ if } z_{p_0} < N_{Di} \le z_{p_0 + p_1} \\ 2 & \text{ otherwise} \end{array}  \right. ,
        } where \eqn{z_k} is the \eqn{k'th} percentile of the standard normal (\eqn{P(Z \le z_k) = k} when \eqn{Z \sim N(0,1)}).
}
\item A binary treatment is simulated as \eqn{X | Z_1,...,Z_n, Z_{D1}, Z_{D2}, U \sim \text{Bernoulli}(p_X)},
where \eqn{p_X = \frac{1}{1 + \exp(\gamma_x + \alpha_U \cdot U + \alpha_1 \cdot Z_1 + ... + \alpha_n \cdot Z_n + \alpha_{D1} \cdot Z_{D1} + \alpha_{D2} \cdot Z_{D2})}},
and all parameters \eqn{\alpha_j} are user specified.
\item A binary outcome is simulated as \eqn{Y | X, U, Z_1,...,Z_n, Z_{D1}, Z_{D2}, U \sim \text{Bernoulli}(p_Y)},
where \deqn{p_Y = \frac{1}{1 + \exp(\gamma_y + \beta_U \cdot U + \beta_X \cdot X + \beta_1 \cdot Z_1 + ... + \beta_n \cdot Z_n + \beta_{D1} \cdot Z_{D1} + \beta_{D2} \cdot Z_{D2})},}
and all parameters \eqn{\beta_j} are user specified.
}
}
