% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_manipulations.R
\name{summary_probs_from_data}
\alias{summary_probs_from_data}
\title{Calculate gammas/thetas from full data}
\usage{
summary_probs_from_data(dat, X = X, Y = Y, Z = Z)
}
\arguments{
\item{dat}{data.frame with columns containing observations of X,Y,Z}

\item{X}{variable with observations on X}

\item{Y}{variable with observations on Y}

\item{Z}{variable with observations on Z}
}
\value{
A list with two vectors: \code{gammas} of length number of different Z values giving P(Y = 1 | Z = z), and \code{thetas} of length number of different Z values giving P(X = 1 | Z = z)
}
\description{
Calculates P(Y = 1 | Z = z) (\code{gammas}) and P(X = 1 | Z = z) (\code{thetas}) from data.frame/tibble with columns X,Y,Z
}
