% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpbounds_misc.R
\name{tidy_bpbounds}
\alias{tidy_bpbounds}
\title{Tidy up output from bpbounds}
\usage{
tidy_bpbounds(bpbs)
}
\arguments{
\item{bpbs}{Output from bpbounds::bpbounds}
}
\description{
Return output from bpbounds::bpbounds in a tidy format.
}
