% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_joint.R
\name{potential_covs}
\alias{potential_covs}
\title{Valid range of Covs}
\usage{
potential_covs(thetas, gammas, x_mono = FALSE)
}
\arguments{
\item{thetas}{vector of length \code{n_z_levels} giving P(X = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}

\item{gammas}{vector of length \code{n_z_levels} giving P(Y = 1 | Z = z), z = 0, ..., (\code{n_z_levels} - 1).}
}
\value{
A list with four elements:
\itemize{
\item \code{thetas} and \code{gammas} provided
\item \code{potential_cov_range} is a tibble with three columns: \code{min} and \code{max} give bounds on Cov(X,Y|Z=z) for the
value of \code{z}
\item \code{constraints}is a tibble with four columns: \code{min} and \code{max} give bounds on Cov(X,Y|Z = z1) - Cov(X,Y | Z = z2) for the
corresponding values of \code{z1} and \code{z2}.
}
}
\description{
From values of P(X = 1 | Z = z) and P(Y = 1 | Z = z) (for all z = 0, ..., \code{n_z_levels} - 1),
this function returns ranges of valid values for the quantities Cov(X,Y|Z = z) and Cov(X,Y|Z = z1) - Cov(X,Y|Z = z2),
such that there probability distribution defined by the input and any set of covariances chosen to fulfill the
constraints will be valid, and not violated the IV inequalities.
}
